requirements.txt
datasets==2.12.0
hydra-core==1.3.2
matplotlib==3.7.1
mlflow==2.3.2
numpy==1.24.3
openai==0.27.8
scikit_learn==1.2.2
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
wandb==0.15.4
xformers==0.0.20
